Migration OWSI-Core : Wicket 6.1.1, Wicket 6.0.0 (premier jet... ça va péter dans tous les sens)
git-svn-id: svn+ssh://svn.projects.openwide.fr/svnroot/owsi-tools/trunk/owsi-core@1607 d3474844-eb8e-4abc-b058-2b321fed648b

+++ owsi-core/owsi-core-components/owsi-core-component-wicket-more/src/main/java/fr/openwide/core/wicket/more/markup/html/navigation/paging/BootstrapPagingNavigation.java
@@ -20,7 +20,7 @@ public class BootstrapPagingNavigation extends PagingNavigation {
 	@Override
 	protected void populateItem(LoopItem loopItem) {
 		super.populateItem(loopItem);
-		int pageIndex = getStartIndex() + loopItem.getIndex();
+		long pageIndex = getStartIndex() + loopItem.getIndex();
 		if (pageable.getCurrentPage() == pageIndex) {
 			loopItem.add(new ClassAttributeAppender("active"));
 		}
@@ -39,7 +39,7 @@ public class BootstrapPagingNavigation extends PagingNavigation {
 	}
 	
 	@Override
-	protected AbstractLink newPagingNavigationLink(String id, IPageable pageable, int pageIndex) {
+	protected AbstractLink newPagingNavigationLink(String id, IPageable pageable, long pageIndex) {
 		return new PagingNavigationLink<Void>(id, pageable, pageIndex).setAutoEnable(false);
 	}
 }
